Refactor query from db_test
