+++ util/testutil/storage.go
@@ -1,0 +1,64 @@
+// Copyright 2016 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package testutil
+
+import (
+	"testing"
+
+	"github.com/prometheus/prometheus/pkg/labels"
+	"github.com/prometheus/prometheus/storage"
+	"github.com/prometheus/prometheus/tsdb/tsdbutil"
+	"github.com/stretchr/testify/require"
+)
+
+type sample struct {
+	t int64
+	v float64
+}
+
+func newSample(t int64, v float64) tsdbutil.Sample { return sample{t, v} }
+func (s sample) T() int64                          { return s.t }
+func (s sample) V() float64                        { return s.v }
+
+// QueryBlock runs a matcher query against the querier and fully expands its data.
+func QueryBlock(t testing.TB, q storage.Querier, matchers ...*labels.Matcher) map[string][]tsdbutil.Sample {
+	ss := q.Select(false, nil, matchers...)
+	defer func() {
+		require.NoError(t, q.Close())
+	}()
+
+	result := map[string][]tsdbutil.Sample{}
+	for ss.Next() {
+		series := ss.At()
+
+		samples := []tsdbutil.Sample{}
+		it := series.Iterator()
+		for it.Next() {
+			t, v := it.At()
+			samples = append(samples, sample{t: t, v: v})
+		}
+		require.NoError(t, it.Err())
+
+		if len(samples) == 0 {
+			continue
+		}
+
+		name := series.Labels().String()
+		result[name] = samples
+	}
+	require.NoError(t, ss.Err())
+	require.Equal(t, 0, len(ss.Warnings()))
+
+	return result
+}
